requirements.txt
pyOpenSSL
texttable~=1.6.3
six~=1.11.0